% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_app}
\alias{run_app}
\title{Run topografie shiny app}
\usage{
run_app(filename = "default.rds", ...)
}
\arguments{
\item{filename}{The file with geo-info to load into the app}

\item{...}{Other parameters to be passed on to shinyApp}
}
\value{
Nothing
}
\description{
Runs a shiny app that displays a map. The user has to provide the correct answer for the item shown on the map.
}
