% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topo_ui.R
\name{topo_ui}
\alias{topo_ui}
\title{Topo-ui}
\usage{
topo_ui(datasets, ...)
}
\arguments{
\item{datasets}{The datasets to choose from}

\item{...}{Not used}
}
\value{
sidebarLayout
}
\description{
Creates a shiny::sidebarLayout to be used in the topo-app
}
